Code review feedback amends
