+++ projects/media-viewer/src/assets/sass/comment-set-panel.scss
@@ -20,7 +20,7 @@
 }
 
 .comments-panel.icp-mode {
-  top: 89px;
+  top: 90px;
   height: 100%
 }
