--- projects/media-viewer/src/lib/toolbar/sub-toolbar/sub-toolbar.component.html
@@ -1,49 +1,0 @@
-<div id="secondaryToolbar" class="secondaryToolbar doorHangerRight" [hidden]="toolbarEvents.subToolbarHidden | async">
-  <div id="secondaryToolbarButtonContainer">
-    <button *ngIf="toolbarButtons.showHighlightButton" (click)="onClickHighlightToggle()"
-      [class.toggled]="toolbarEvents.highlightModeSubject | async" class="secondaryToolbarButton highlightBtn"
-      title="Highlight Text Mode" id="secondaryHighlightToggle" tabindex="0" data-l10n-id="secondaryHighlightToggle">
-      <span data-l10n-id="highlight_label">Highlight Text</span>
-    </button>
-    <button *ngIf="toolbarButtons.showDrawButton" (click)="onClickDrawToggle()"
-      [class.toggled]="toolbarEvents.drawModeSubject | async" class="secondaryToolbarButton drawBtn" title="Draw Mode"
-      id="secondaryToggleDrawButton" tabindex="1" data-l10n-id="secondaryToggleDrawButton">
-      <span data-l10n-id="draw_label">Draw</span>
-    </button>
-    <div class="horizontalToolbarSeparator"></div>
-    <button *ngIf="toolbarButtons.showRotate" class="secondaryToolbarButton rotateCcw" title="Rotate Counterclockwise"
-      tabindex="4" data-l10n-id="page_rotate_ccw" (click)="rotateCcw()" id="secondaryRotateCcw">
-      <span data-l10n-id="page_rotate_ccw_label">Rotate Counterclockwise</span>
-    </button>
-    <button *ngIf="toolbarButtons.showRotate" class="secondaryToolbarButton rotateCw" title="Rotate Clockwise"
-      tabindex="5" id="secondaryRotateCw" data-l10n-id="page_rotate_cw" (click)="rotateCw()">
-      <span data-l10n-id="page_rotate_cw_label">Rotate Clockwise</span>
-    </button>
-    <button *ngIf="toolbarButtons.showGrabNDragButton" [class.toggled]="toolbarEvents.grabNDrag | async"
-      class="secondaryToolbarButton grabNDrag" title="Grab N Drag" tabindex="6" id="grabNDrag"
-      data-l10n-id="page_grab_n_drag" (click)="grabNDrag()">
-      <span data-l10n-id="page_grab_n_drag_label">Grab N Drag</span>
-    </button>
-    <button *ngIf="toolbarButtons.showSaveRotationButton" class="secondaryToolbarButton saveRotation"
-      title="Save Rotation" tabindex="7" id="saveRotation" data-l10n-id="page_save_rotation" (click)="saveRotation()">
-      <span data-l10n-id="page_save_rotation_label">Save Rotation</span>
-    </button>
-    <div class="horizontalToolbarSeparator"></div>
-    <button id="secondaryPrint" *ngIf="toolbarButtons.showPrint" class="secondaryToolbarButton print" title="Print"
-      tabindex="2" data-l10n-id="print" (click)="printFile()">
-      <span data-l10n-id="print_label">Print</span>
-    </button>
-    <button id="secondaryDownload" *ngIf="toolbarButtons.showDownload" class="secondaryToolbarButton download"
-      title="Download" tabindex="3" data-l10n-id="download" (click)="downloadFile()">
-      <span data-l10n-id="download_label">Download</span>
-    </button>
-    <div class="horizontalToolbarSeparator"></div>
-    <div *ngIf="!icpEnabled">
-      <button id="icpCreateSession" class="secondaryToolbarButton" title="IcpCreateSession"
-        data-l10n-id="create_icp_session" (click)="togglePresentBar()">
-        <span data-l10n-id="create_icp_session">Launch Presentation</span>
-      </button>
-    </div>
-    <div class="horizontalToolbarSeparator"></div>
-  </div>
-</div>
